+++ node/immutable-store.js
@@ -1,12 +1,4 @@
 "use strict";
-var __assign = (this && this.__assign) || Object.assign || function(t) {
-    for (var s, i = 1, n = arguments.length; i < n; i++) {
-        s = arguments[i];
-        for (var p in s) if (Object.prototype.hasOwnProperty.call(s, p))
-            t[p] = s[p];
-    }
-    return t;
-};
 var index_1 = require("./index");
 var typescript_promise_1 = require("typescript-promise");
 var _filters = {
@@ -495,14 +487,27 @@ var _SomeSQLQuery = (function () {
                 callBack([{ msg: "0 rows " + describe }]);
             }
         };
+        var freezeObj = function (obj) {
+            for (var key in obj) {
+                if (obj.hasOwnProperty(key) && obj[key] instanceof Object) {
+                    obj[key] = freezeObj(obj[key]);
+                }
+            }
+            return Object.freeze(obj);
+        };
         var updateRow = function (rowID, cb) {
             changedRowIDs.push(rowID);
-            var newRow = __assign({}, t._db._getRow(rowID) || {});
-            // let newRow = JSON.parse(JSON.stringify(t._getRow(rowID) || {}));
+            var newRow = JSON.parse(JSON.stringify(t._db._getRow(rowID) || {}));
             for (var key in qArgs) {
                 newRow[key] = cb(key, newRow[key]);
             }
-            t._db._rows[rowID].unshift(Object.freeze(newRow));
+            // Add default values if the value is null;
+            t._db._models[t._db._selectedTable].forEach(function (model) {
+                if (model.default && !newRow[model.key]) {
+                    newRow[model.key] = model.default;
+                }
+            });
+            t._db._rows[rowID].unshift(freezeObj(newRow));
             if (t._db._indexedDB) {
                 var tableName = t._db._tableInfo[t._db._selectedTable]._name;
                 t._db._indexedDB.transaction(tableName, "readwrite").objectStore(tableName).put(newRow);